+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.8
 * General:
     * Supports more markdown syntax (strikethrough, tables, simple links)
+* PDF:
+    * Fix of crash on eLearnJS interface functions
 ## 0.2.7
 * General:
     * Selection of extensions whether to include them in an export or not